Keep shop item data persistent for shops with a reset timer
